layout corrected, filter not working
